+++ seata-storage/src/main/java/com/seata/storage/service/StorageService.java
@@ -1,5 +1,5 @@
 package com.seata.storage.service;
 
 public interface StorageService {
-    public void decrease(Integer productId,Integer count);
+    public void decrease(Long productId,Integer count);
 }
